makes different sub-folder in quiz for question specific code
