+++ backend/src/api/quizes/[id]/questions/[questionID].ts
@@ -4,7 +4,8 @@ import { tryAndCatch } from '@lib/middleware';
 import { QuizModel } from '@lib/quiz/quiz.model';
 import { ensureAdmin } from '@lib/session/session.middleware';
 import { validateUsing } from '@lib/validator/joi.middleware';
-import { updateQuestionvalidate } from '@lib/quiz/quiz.validator';
+import { updateQuestionValidate } from '@lib/quiz/question/question.validator';
+
 
 const updateQuestion: Handler = async (req: Request, res: Response) => {
    let updatedQuiz = await QuizModel.findOneAndUpdate(
@@ -16,5 +17,5 @@ const updateQuestion: Handler = async (req: Request, res: Response) => {
    res.status(200).json(updatedQuiz);
 };
 
-export const put: Handler[] = [ensureAdmin, validateUsing(updateQuestionvalidate), tryAndCatch(updateQuestion)];
+export const put: Handler[] = [ensureAdmin, validateUsing(updateQuestionValidate), tryAndCatch(updateQuestion)];
